improve merger

+++ Batch.roboFontExt/lib/binaryMerger/__init__.py
@@ -9,6 +9,7 @@ from fontTools.ttLib import TTFont
 
 from mojo.extensions import getExtensionDefault, setExtensionDefault
 from mojo.roboFont import OpenFont
+from mojo.UI import getDefault
 
 from batchTools import settingsIdentifier, Report
 
@@ -76,7 +77,10 @@ class BinaryMerger(Group):
         paths = self.controller.get(["ufo"])
 
         report = Report()
-        tempDir = tempfile.mkdtemp()
+
+        tempDir = os.path.join(destDir, "temp")
+        if not os.path.exists(tempDir):
+            os.makedirs(tempDir)
         tempExportPaths = self._generateCallback(tempDir, progress, report)
 
         progress.update("Merging Tables...")
@@ -93,33 +97,36 @@ class BinaryMerger(Group):
                 binaryIsOpenType = os.path.splitext(binarySourcepath)[1].lower() in [".ttf", ".otf"]
                 tempIsOpenType = os.path.splitext(tempExportPath)[1].lower() in [".ttf", ".otf"]
                 if binaryIsOpenType and tempIsOpenType:
-                    binarySource = TTFont(binarySourcepath)
-                    tempFont = TTFont(tempExportPath)
-                    fileName = os.path.basename(tempExportPath)
-                    if not self.controller.keepFileNames():
-                        fileName = "%s-%s" % (font.info.familyName, font.info.styleName)
-                    path = os.path.join(destDir, fileName)
-                    report.writeTitle(os.path.basename(path), "'")
-                    report.write("source: %s" % path)
-                    report.write("binary source: %s" % binarySource)
-                    report.newLine()
-                    report.indent()
-                    for table in tableNames:
-                        if table in binarySource:
-                            report.write("merge %s table" % table)
-                            tempFont[table] = binarySource[table]
-                    report.dedent()
-                    report.newLine()
-                    tempFont.save(path)
-                    tempFont.close()
-                    binarySource.close()
+                    if os.path.exists(binarySourcepath) and os.path.exists(tempExportPath):
+                        binarySource = TTFont(binarySourcepath)
+                        tempFont = TTFont(tempExportPath)
+                        fileName = os.path.basename(tempExportPath)
+                        if not self.controller.keepFileNames():
+                            fileName = "%s-%s%s" % (font.info.familyName, font.info.styleName, os.path.splitext(tempExportPath)[1])
+                        path = os.path.join(destDir, fileName)
+                        report.writeTitle(os.path.basename(path), "'")
+                        report.write("source: %s" % tempExportPath)
+                        report.write("binary source: %s" % binarySourcepath)
+                        report.newLine()
+                        report.indent()
+                        for table in tableNames:
+                            if table in binarySource:
+                                report.write("merge %s table" % table)
+                                tempFont[table] = binarySource[table]
+                        report.write("save to %s" % path)
+                        tempFont.save(path)
+                        report.dedent()
+                        report.newLine()
+                        tempFont.close()
+                        binarySource.close()
             font.close()
 
         reportPath = os.path.join(destDir, "Binary Merge Report.txt")
         report.save(reportPath)
 
-        if os.path.exists(tempDir):
-            shutil.rmtree(tempDir)
+        if not getDefault("Batch.Debug", False):
+            if os.path.exists(tempDir):
+                shutil.rmtree(tempDir)
 
     def _generate(self, destDir):
         if not destDir:
